Updated master display coordinates to be decimals.

+++ Tricycle.Models/Media/MasterDisplayProperties.cs
@@ -2,10 +2,10 @@
 {
     public class MasterDisplayProperties
     {
-        public Coordinate<int> Red { get; set; }
-        public Coordinate<int> Green { get; set; }
-        public Coordinate<int> Blue { get; set; }
-        public Coordinate<int> WhitePoint { get; set; }
-        public Range<int> Luminance { get; set; }
+        public Coordinate<decimal> Red { get; set; }
+        public Coordinate<decimal> Green { get; set; }
+        public Coordinate<decimal> Blue { get; set; }
+        public Coordinate<decimal> WhitePoint { get; set; }
+        public Range<decimal> Luminance { get; set; }
     }
 }
